Initial commit
Right now, the codebase constitutes a small framework (emph on frame and
not work) for a PoC socket-based first-principles admm solution.

Constituents:
1. Pyproject toml with the project information, dependencies and
   parameters for automated CI/CD.
2. Flake files for reproducible atomic hydra builds.
3. README stub.
4. Network layer classes in _net.py which mainly exist for the purposes
   of server-client(s) numpy array sharing.
5. Argument parser for the final application in _parser.py.
6. Type aliases and generics to be used throughout the project in
   _types.py.
7. Base admm implementation in admm.py.
8. Lasso admm implementation in lasso_admm.py.

N.B: the structure of this project is one of a _python package_, which
has implications for relative imports (i.e. import a v. import .a).
There's still quite a bit to do in order to make everything semantically
and functionally operational, but this seems like a good start.