(feat) Gaussian RDP noise
Implements a basic noise scaler parametrized on yielding (α,ε)-RDP. See
the Cyffers, Dwork and the two (2) Mironov papers for more details. The
base `ADMM` class uses the `get_noise_scale` function to get a noise
standard deviation which is then directly applied to numpy's
`random.normal` with `scale=get_noise_scale(...)`. How exactly this
function is implemented is relegated to each child objective class.

In order to implement this function for some objective, there are two
matters one has to consider:
1. Establishing the lp-sensitivity of the objective i.e. how much the lp
   norm of the objective shifts when using adjacent datasets.
2. Establishing a correspondance between characteristics of the
   objective and the privacy budget ε.

For the purposes of this assignment, we use the RDP formulation with the
Rényi divergence. However, it should be possible to express ε with any
DP framework given enough patience.