(form)(opt) generic method getter
In many parts of the codebase we "get" a callable by indexing some
dict with a given key (e.g. getting the `_l_mae` loss function with the
key `mae`). This is done using a dict associating a public external name
(e.g. "mae") with a private internal function (e.g. `_l_mae`) and a
"getter" function which searches this dict and fails gracefully if the
query returns `KeyError`.

This paradigm is used multiple times within the codebase which lead to
repetitive definitions of this getter function.

This commit converts the different getters into essentially "stubs"
which call a generic getter `call_gettr` in `_utils.py`. `call_gettr`
takes two arguments:
  1. The key to look for.
  2. The dict to look into.
and replicates the graceful fallback behaviour that was previously
implemented in each getter.