(opt) create an abstract formulation for dp noise
This formulation splits the noise generating function into two
components:
  1. Sensitivity
  2. Mechanism
This allows generating noise for any combination of randomized mechanism
and function which is extremely useful if we need to apply DP in
multiple _different_ queries. It also allows one to specify the noise
generating mechanism (e.g. Gaussian, Laplacian, etc) to be used for
all DP processes.