(feat) replace loss hist with coeff hist
Explicitly computing the loss doesn't make much sense within the context
of the ADMM classes. That is better left for the user (cf. `sklearn`).

This commit replaces the loss history mechanism with a simple recording
of all coefficients of all clients + the server for all iterations. The
shape of the resulting `coeff_hist` array should be:
  (№ clients, № iterations, № features)
Due to its 3d shape, the array cannot be stored in plaintext and
as such, `npy` files are used in place of the more conventional `csv`.

This approach significantly simplifies the code.