_args::run_server: force coeff line export
By default, if the input is a vector (shape : (n,)), `savetxt` exports
to one column. In order to make manipulation of the csv file easier
(e.g. appending to other results, searching by column, etc), this commit
forces the exported coeffs to a 2d array and exports that instead. This
way, the generated `csv` file is always a _line_ instead of a _column.